Backend/testing/requirements.txt
numpy==2.1.3
pytest==8.3.3
pytest-asyncio==0.24.0
python-dotenv==1.0.1
requests==2.32.3
websockets==14.1